Refactor. New Popouts
Now using inheritance and abstraction
