+++ SpeedrunNotes/Popouts/BasePopout.xaml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             x:Class="SpeedrunNotes.Popouts.BasePopout">
+</ContentPage>
