+++ SpeedrunNotes/Popouts/NextSplitPopout.xaml
@@ -1,8 +1,9 @@
 <?xml version="1.0" encoding="utf-8" ?>
-<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="SpeedrunNotes.Popouts.NextSplitPopout"
-             Title="NextSplitPopout" >
+<local:BasePopout xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+              xmlns:local="clr-namespace:SpeedrunNotes.Popouts"
+              x:Class="SpeedrunNotes.Popouts.NextSplitPopout"
+              Title="NextSplitPopout" >
 
     <Grid>
         <!-- Next split -->
@@ -19,4 +20,4 @@
                    TextColor="White" />
         </Grid>
     </Grid>
-</ContentPage>
+</local:BasePopout>
