+++ SpeedrunNotes/Popouts/SplitNote2Popout.xaml
@@ -1,10 +1,11 @@
 <?xml version="1.0" encoding="utf-8" ?>
-<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+<local:BasePopout xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+              xmlns:local="clr-namespace:SpeedrunNotes.Popouts"
              x:Class="SpeedrunNotes.Popouts.SplitNote2Popout"
              Title="SplitNote1Popout" >
 
     <Grid>
         <Label x:Name="SplitNoteLabel" />
     </Grid>
-</ContentPage>
+</local:BasePopout>
