+++ SpeedrunNotes/Popouts/SplitNote2Popout.xaml.cs
@@ -2,7 +2,7 @@ using CommunityToolkit.Mvvm.Messaging;
 
 namespace SpeedrunNotes.Popouts;
 
-public partial class SplitNote2Popout : ContentPage
+public partial class SplitNote2Popout : BasePopout
 {
     int FontSize;
     string SplitNoteLabelText;
@@ -24,29 +24,7 @@ public partial class SplitNote2Popout : ContentPage
         InitTimer();
     }
 
-    protected override void OnAppearing()
-    {
-        base.OnAppearing();
-
-        Window.MinimumWidth = 320;
-        Window.MinimumHeight = 200;
-    }
-
-    public void InitTimer()
-    {
-        // Setup timer to send / recieve message with LiveSplit.Server every second
-        System.Timers.Timer Timer = new(1000);
-        Timer.Elapsed += OnTimedEvent;
-        Timer.AutoReset = true;
-        Timer.Enabled = true;
-    }
-
-    void OnTimedEvent(object sender, EventArgs e)
-    {
-        MainThread.BeginInvokeOnMainThread(UpdateNextSplit);
-    }
-
-    void UpdateNextSplit()
+    public override void UpdateData()
     {
         SplitNoteLabel.Text = SplitNoteLabelText;
         SplitNoteLabel.FontSize = FontSize;
